+++ packages/dumb/src/models/enums/bolus-type.enum.ts
@@ -28,7 +28,7 @@
 export enum BolusType {
   Manual = 'Manual',
   Meal = 'Meal',
-  Micro = 'Micro',
+  Correction = 'Correction',
   Pen = 'Pen',
   Umm = 'Umm'
 }