+++ CHANGELOG.md
@@ -17,6 +17,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-3087 As a user, I want to see the medisafe pump event in yourloops
 - YLP-3141 As a user, I want my alarms messages to be the same as my device's
 - YLP-3165 Update basal color
+- YLP-3177 As a user, I want to see that "micro boluses" are now called "correction boluses"
 
 ### Fixed
 - YLP-2632 Data is uploaded with a date in the future